+++ src/D20Tek.BlazorComponents.Toast/DependencyInjection.cs
@@ -2,6 +2,12 @@ namespace D20Tek.BlazorComponents;
 
 public static class DependencyInjection
 {
-    public static IServiceCollection AddToast(this IServiceCollection services) =>
-        services.AddScoped<IToastService, ToastService>();
+    public static IServiceCollection AddToast(this IServiceCollection services, Action<ToastDefaults>? configure = null)
+    {
+        var defaults = new ToastDefaults();
+        configure?.Invoke(defaults);
+
+        return services.AddSingleton(defaults)
+                       .AddScoped<IToastService, ToastService>();
+    }
 }