+++ src/D20Tek.BlazorComponents.Toast/ToastDefaults.cs
@@ -1,0 +1,14 @@
+namespace D20Tek.BlazorComponents;
+
+public sealed class ToastDefaults
+{
+    public ToastPosition Position { get; set; } = ToastPosition.BottomRight;
+
+    public TimeSpan DefaultTimeout { get; set; } = TimeSpan.FromSeconds(3);
+
+    public bool ShowIcon { get; set; } = true;
+
+    public bool Dismissible { get; set; } = true;
+
+    public bool Animate { get; set; } = true;
+}